requirements.txt
astroid==1.5.3
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
PyJWT==1.5.3
pylint==1.7.4
pytz==2017.3
six==1.11.0
wrapt==1.10.11
